enterprise_code_backup/addons/aicleaner_v3/requirements.txt
# AICleaner v3 Requirements
# Core dependencies for Home Assistant addon with Phase 3A+3B features

# AI/ML Libraries
google-generativeai>=0.3.0
anthropic>=0.7.0
openai>=1.0.0
ollama>=0.1.7

# Image Processing
Pillow>=10.0.0

# System Monitoring
psutil>=5.9.0

# Async HTTP Client
aiohttp>=3.8.0

# MQTT Integration
aio-mqtt>=0.15.0

# Testing Framework
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0

# Data Handling
pydantic>=2.0.0
PyYAML>=6.0

# API Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
slowapi>=0.1.9

# Utilities
python-dateutil>=2.8.0
typing-extensions>=4.0.0

# Home Assistant Integration
homeassistant>=2023.1.0

# Optional dependencies for enhanced functionality
# These are marked as optional but recommended for full feature set
requests>=2.28.0
urllib3>=1.26.0
certifi>=2022.0.0

# Development dependencies (optional)
black>=22.0.0
flake8>=5.0.0
mypy>=1.0.0